Fixed error with preview image when dragging.
Fixed refresh after moving or resizing.
Fixed boundary and bounding box not being updated after changing scale.

The problem was due to how, when, and why each shape is drawn.
The preview shape is created at each new location of the 'drag' path - scale will not dynamically change.
The thumbnail shape is the scaled down/up image created for the tool bar - only created when the tool bar needs to be redrawn - scale changes to conform to the tool bar's size.
The display shape is the shape that is used by the user and the scaling values will persist for this shape - scale can be changed at any time.

In order to fix the problem I needed to alter some of the properties and add a few.
The class width and height are floating point values and allow the representation of the shape in floating precision.
However, NShape uses int for width and height - which doesn't allow for the floating point precision.
To implement floating precision:
The width values passed to the base class had to be multiplied by the factor value and the pixel to millimeter value, then rounding the result up to the nearest int.
This is used by the underlying NShape system to handle the shape logic (boundaries and the such).
Changed the previous float properties to WidthF and HeightF so they are not hiding inherited members.
Had to create a new Width and Height properties with an int return value.
Set them as private in order to keep them from being visible to derived classes and property grid.
These are dynamically changed when the scale value is changed.
The WidthF and HeightF always remain the same.
They are used each time when drawing the shape and are calculated against the scale and pixel to millimeter value and are drawn to floating precision by the underlying GDI system that NShape exposes.

Summary:
NShapes Width and Height can always change (required for visual effects of the library).
The actual shapes WidthF and HeightF values never change (an RJ45 will always be the same width and height).
The visual shape that is seen is a only a scaled floating point visual representation.

Should add an option in tool for batch loading and processing in future.